Reorganization is mostly complete.
Size/scaling has been normalized for all objects and is now consolidated
inside each object.
The tank, environment, and entities are all encapsulated and organized
within a Scenario object.
Still some work to be done on saving & loading for Scenarios, but that
shouldn't require moving anything else around.
Currently, the environment is empty to start. Test entities can be added
with the following hotkeys:
	K: add fish
	L: add plant
	P: add pot
The locations are currently fixed, but the objects can be viewed with
the proper scale.