Yay actual visible functionality!
Now, if you click a fish, plant, or pot (excluding the player fish that
starts in the middle) it will be selected and glow. If you click a
different object selection should switch to that object. If you hold
down ctrl and click one that is already selected it should deselect it.