Selected Entities can now be moved.
Click the "Move" button in the bottom right and select an entity. Move
left, right, forward, backward with WASD. Fish can be moved up and down
as well with the up and down arrows. Note that fish movement is a bit
weird since they are already moving around, but it does still work.